lyra-backend/requirements.txt
Flask
spotipy
python-dotenv
requests
Flask-Cors
google-generativeai
waitress
